requirements.txt
# Core dependencies
openai==1.82.1
azure-identity>=1.17.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Token counting and embeddings
tiktoken==0.5.2
numpy==1.26.3
scikit-learn==1.4.0

# Repository management
GitPython==3.1.41

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3

# Web requests
requests==2.31.0

# Progress bars
rich

# Optional but recommended
tenacity>=8.2.0      # For retry logic in batch embedding processing (handles rate limits gracefully)

# Development tools
pytest  # Testing framework
black  # Code formatting
isort  # Import sorting 